🧹 Complete canvas refactor cleanup
- Delete old 907-line CanvasArea.tsx monolith
- Rename CanvasAreaRefactored to CanvasArea
- Update all imports to use new component
- Remove comparison page (no longer needed)
- Update migration checklist as COMPLETE

"The old canvas has been overcome. The Übercomponent reigns supreme\!"
