+++ app/page.tsx
@@ -6,8 +6,7 @@ import { exportCanvasAsPNG } from '@/lib/export-utils'
 import { SaveDialog } from '@/components/save-dialog'
 import { SavedItemsDialog } from '@/components/saved-items-dialog'
 import { StudioHeader } from '@/components/studio/StudioHeader'
-// import { CanvasArea } from '@/components/studio/CanvasArea'
-import { CanvasAreaRefactored as CanvasArea } from '@/components/studio/CanvasAreaRefactored'
+import { CanvasArea } from '@/components/studio/CanvasArea'
 import { ControlsPanel } from '@/components/studio/ControlsPanel'
 import { BrandPresetsPanel } from '@/components/studio/BrandPresetsPanel'
 import { IndustrySelector } from '@/components/studio/IndustrySelector'
