+++ docs/canvas-refactor-migration.md
@@ -25,12 +25,13 @@ Refactoring the 907-line `CanvasArea.tsx` monolith into a modular, maintainable 
 - [ ] Check preview mode functionality
 - [ ] Test export functionality
 
-### Phase 3: Migration
-- [ ] Update imports in `app/page.tsx`
-- [ ] Replace `CanvasArea` with `CanvasAreaRefactored`
-- [ ] Update any dependent components
-- [ ] Remove old `CanvasArea.tsx`
-- [ ] Update tests if any exist
+### Phase 3: Migration ✅
+- [x] Update imports in `app/page.tsx`
+- [x] Replace `CanvasArea` with `CanvasAreaRefactored`
+- [x] Update any dependent components
+- [x] Remove old `CanvasArea.tsx`
+- [x] Rename refactored component to original name
+- [x] Clean up test pages
 
 ### Phase 4: Optimization
 - [ ] Add error boundaries
